requirements.txt
numpy
theseus-ai
tqdm
configargparse
h5py
pyglet
trimesh
matplotlib
torchvision
tensorboard